typeck/expr.rs: move check_method_call here.
